requirements.txt
ansible==2.3
google-cloud-storage==1.8.0